src/requirements.txt
fastapi==0.111.0
pydantic==2.8.2
pymongo==4.8.0
uvicorn==0.30.1
paho-mqtt==1.6.1